fix4

+++ tasks/mpi/sedova_o_max_of_vector_elements/perf_tests/main.cpp
@@ -25,9 +25,9 @@ std::vector<std::vector<int>> generate_random_matrix(size_t rows, size_t cols, s
 }
 
 TEST(sedova_o_max_of_vector_elements_mpi, test_pipeline_run) {
-  int rows = 7000;
-  int cols = 7000;
-  int value = 7000;
+  size_t rows = 7000;
+  size_t cols = 7000;
+  size_t value = 7000;
 
   boost::mpi::communicator world;
   std::vector<std::vector<int>> global_matrix;
@@ -78,9 +78,9 @@ TEST(sedova_o_max_of_vector_elements_mpi, test_pipeline_run) {
 }
 
 TEST(sedova_o_max_of_vector_elements_mpi, test_task_run) {
-  int rows = 7000;
-  int cols = 7000;
-  int value = 7000;
+  size_t rows = 7000;
+  size_t cols = 7000;
+  size_t value = 7000;
 
   boost::mpi::communicator world;
   std::vector<std::vector<int>> global_matrix;
